Update to screen.css
This simple edit adds submenu support to the Pluto Wordpress theme. It contains the correct code to have your submenus appear on hover and become clickable. It also includes a visibility:hidden parameter to hide the corner decorations of the slide-out pages. Just delete lines 1647, 1656, 1665, and 1674 if you like the page corner decorations.

To see a sample of these modifications, go to http://jetbalagtas.com/portfolio/bushido/